# Application/Frigorino.Web/ClientApp/src/lib/api/models/CreateInventoryRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateInventoryRequest = {
    name?: string | null;
    description?: string | null;
};