# Application/Frigorino.Web/ClientApp/src/lib/api/models/UpdateInventoryRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdateInventoryRequest = {
    name?: string | null;
    description?: string | null;
};